pyproject.toml
[tool.poetry]
name = "premier-geek"
version = "0.1.0"
description = "AI-powered Premier League squad intelligence"
authors = ["Sebastian Kucharzyk <kucharzyk.sebastian@gmail.com>"]
readme = "README.md"
packages = [{include = "premier_geek"}]

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
aws-cdk-lib = "^2.154.1"
constructs = "^10.3.0"
cdk-poetry-asset-code = "^0.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[tool.poetry.group.typing.dependencies]
pyright = "^1.1.377"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
fix = true
line-length = 120
src = ["premier-geek"]
target-version = "py312"

[tool.ruff.lint]
extend-select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ERA", # flake8-eradicate/eradicate
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PGH", # pygrep
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
]
extend-safe-fixes = [
    "TCH", # move import from and to TYPE_CHECKING blocks
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2

[tool.pyright]
reportMissingImports = true
pythonVersion = "3.12"
pythonPlatform = "Linux"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
